requirements.txt
PyMuPDF==1.23.26
python-docx==1.1.2
google-cloud-aiplatform==1.55.0
google-generativeai==0.7.1
google-cloud-texttospeech==2.16.2
pydub==0.25.1
EbookLib==0.18
mobi==0.3
beautifulsoup4==4.12.3

requests==2.32.3
tqdm==4.66.4
tenacity==8.2.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
spacy==3.7.2
